Add support for apache-style file directories.
This patch (against version 0.9) allows curlftpfs to
work with web
sites that serve apache-style file directories.

It adds a new directory format parser in ftpfs-ls.c,
which handles
apache-style format listings. For this reason, the
file sizes will
not be exact, and file modification times may be off by
a hour due to
daylight savings. Files and directories as listed in
the index are
handled properly. Web sites which do not serve file
indices are also
handled properly (i.e. are ignored).

Very few changes are required to the code to handle the
HTTP protocol
(other than a few setup details), since libcurl handles
the different
protocols transparently.