Allow for specifying path to netrc
This allows users to give the location to the netrc file, which often
contains secrets for authentication into an FTP server.

Typically, this file lives under `$HOME/.netrc`. However, in certain
scenarios it is desirable to keep this file somewhere other than the
user's directory.

This is basically the same patch from
https://sourceforge.net/p/curlftpfs/discussion/542749/thread/3becfed6/
which was never merged upstream.